chore: move brave rewards information to base cache
