+++ components/brave_wallet_ui/common/async/base-query-cache.ts
@@ -18,7 +18,8 @@ import {
   SupportedTestNetworks,
   SupportedOnRampNetworks,
   SupportedOffRampNetworks,
-  ERC721Metadata
+  ERC721Metadata,
+  BraveRewardsInfo
 } from '../../constants/types'
 
 // entities
@@ -49,6 +50,17 @@ import { addLogoToToken } from './lib'
 import { makeNetworkAsset } from '../../options/asset-options'
 import { isIpfs } from '../../utils/string-utils'
 import { getEnabledCoinTypes } from '../../utils/api-utils'
+import {
+  BraveRewardsProxy,
+  WalletStatus,
+  getBraveRewardsProxy
+} from './brave_rewards_api_proxy'
+import {
+  getRewardsBATToken,
+  getNormalizedExternalRewardsWallet,
+  getNormalizedExternalRewardsNetwork,
+  getRewardsProviderName
+} from '../../utils/rewards_utils'
 
 /**
  * A function to return the ref to either the main api proxy, or a mocked proxy
@@ -70,6 +82,22 @@ export const setApiProxyFetcher = (fetcher: () => WalletApiProxy) => {
 }
 
 /**
+ * A function to return the ref to either the main api proxy, or a mocked proxy
+ * @returns function that returns an ApiProxy instance
+ */
+export let rewardsProxyFetcher = getBraveRewardsProxy
+
+/**
+ * Assigns a function to use for fetching a BraveRewardsProxy
+ * (useful for injecting spies during testing)
+ * @param fetcher A function to return the ref to either the main api proxy,
+ *  or a mocked proxy
+ */
+export const setRewardsProxyFetcher = (fetcher: () => BraveRewardsProxy) => {
+  rewardsProxyFetcher = fetcher
+}
+
+/**
  * A place to store & manage dependency data for other queries
  */
 export class BaseQueryCache {
@@ -82,6 +110,7 @@ export class BaseQueryCache {
   private _extractedIPFSUrlRegistry: Record<string, string | undefined> = {}
   private _enabledCoinTypes: number[]
   private _erc721MetadataRegistry: Record<string, ERC721Metadata>
+  public rewardsInfo: BraveRewardsInfo | undefined = undefined
 
   getWalletInfo = async () => {
     if (!this.walletInfo) {
@@ -350,6 +379,49 @@ export class BaseQueryCache {
 
     return this._erc721MetadataRegistry[tokenId]
   }
+
+  // Brave Rewards
+  getBraveRewardsInfo = async () => {
+    if (!this.rewardsInfo) {
+      const isRewardsEnabled = await rewardsProxyFetcher().getRewardsEnabled()
+
+      if (!isRewardsEnabled) {
+        this.rewardsInfo = emptyRewardsInfo
+        return this.rewardsInfo
+      }
+
+      const balance = await rewardsProxyFetcher().fetchBalance()
+
+      const { provider, status, links } =
+        (await rewardsProxyFetcher().getExternalWallet()) || {}
+
+      const rewardsToken = getRewardsBATToken(provider)
+
+      const isConnected = status === WalletStatus.kConnected
+
+      const rewardsAccount = isConnected
+        ? getNormalizedExternalRewardsWallet(provider)
+        : undefined
+
+      const rewardsNetwork = isConnected
+        ? getNormalizedExternalRewardsNetwork(provider)
+        : undefined
+
+      this.rewardsInfo = {
+        isRewardsEnabled: true,
+        balance,
+        provider,
+        rewardsToken,
+        status: status || WalletStatus.kNotConnected,
+        rewardsAccount,
+        rewardsNetwork,
+        accountLink: links?.account,
+        providerName: getRewardsProviderName(provider)
+      }
+    }
+
+    return this.rewardsInfo
+  }
 }
 
 let cache = new BaseQueryCache()
@@ -586,3 +658,16 @@ export async function makeTokensRegistry(
   )
   return userTokensByChainIdRegistry
 }
+
+// defaults
+export const emptyRewardsInfo: BraveRewardsInfo = {
+  isRewardsEnabled: false,
+  balance: undefined,
+  rewardsToken: undefined,
+  provider: undefined,
+  providerName: '',
+  status: WalletStatus.kNotConnected,
+  rewardsAccount: undefined,
+  rewardsNetwork: undefined,
+  accountLink: undefined
+} as const
