+++ components/brave_wallet_ui/components/desktop/popup-modals/filter-modals/filter-components/filter-networks-section.tsx
@@ -6,9 +6,7 @@
 import * as React from 'react'
 
 // Constants
-import {
-  emptyRewardsInfo //
-} from '../../../../../common/slices/endpoints/rewards.endpoints'
+import { emptyRewardsInfo } from '../../../../../common/async/base-query-cache'
 
 // Queries
 import {
